Do not generate COND SETF bindings for invalid places
COND threading macros no longer generate code that attempts to set
invalid places, such as literals or lists whose first symbol is from
the CL package but is not defined as naming a valid place.

The heuristic is imperfect, but should be enough to satisfy Common Lisp
compilers and not generate compilation warnings anymore.